+++ docs/search/all_15.js
@@ -2,5 +2,5 @@ var searchData=
 [
   ['validateconnection',['ValidateConnection',['../class_pip_services3_1_1_components_1_1_connect_1_1_composite_connection_resolver.html#ad28f5e7477f8fac4b4ce71ae47bd7680',1,'PipServices3::Components::Connect::CompositeConnectionResolver']]],
   ['validatecredential',['ValidateCredential',['../class_pip_services3_1_1_components_1_1_connect_1_1_composite_connection_resolver.html#a7b5f9a455c01e87ddf04ea9775774a3b',1,'PipServices3::Components::Connect::CompositeConnectionResolver']]],
-  ['value',['Value',['../class_pip_services3_1_1_components_1_1_cache_1_1_cache_entry.html#ac5a1e4897fa08665e6d953aa10feae49',1,'PipServices3::Components::Cache::CacheEntry']]]
+  ['value',['Value',['../class_pip_services3_1_1_components_1_1_cache_1_1_cache_entry.html#ac5a1e4897fa08665e6d953aa10feae49',1,'PipServices3.Components.Cache.CacheEntry.Value()'],['../class_pip_services3_1_1_components_1_1_state_1_1_state_value.html#a97d27623761832e9b7a856839f778d4d',1,'PipServices3.Components.State.StateValue.Value()']]]
 ];