+++ docs/search/all_7.js
@@ -2,8 +2,11 @@ var searchData=
 [
   ['get',['Get',['../class_pip_services3_1_1_components_1_1_count_1_1_cached_counters.html#a576ccb51b553133fd3b86994e47bb0a0',1,'PipServices3::Components::Count::CachedCounters']]],
   ['getall',['GetAll',['../class_pip_services3_1_1_components_1_1_auth_1_1_credential_resolver.html#a06b4e2200b9ccf4b019df8ddadab5aa3',1,'PipServices3.Components.Auth.CredentialResolver.GetAll()'],['../class_pip_services3_1_1_components_1_1_connect_1_1_connection_resolver.html#a63f52acc2d6e5f24785fd9443dac772a',1,'PipServices3.Components.Connect.ConnectionResolver.GetAll()'],['../class_pip_services3_1_1_components_1_1_count_1_1_cached_counters.html#a4b276e9a5aeaeac48ba559c388e9d24e',1,'PipServices3.Components.Count.CachedCounters.GetAll()']]],
+  ['getkey',['GetKey',['../class_pip_services3_1_1_components_1_1_state_1_1_state_entry.html#acce608074768e13e24e416c0b8b1389d',1,'PipServices3::Components::State::StateEntry']]],
+  ['getlastupdatetime',['GetLastUpdateTime',['../class_pip_services3_1_1_components_1_1_state_1_1_state_entry.html#ad0028319935ecfb655b7990134df6e7f',1,'PipServices3::Components::State::StateEntry']]],
   ['getport',['GetPort',['../class_pip_services3_1_1_components_1_1_connect_1_1_connection_params.html#a8a332ae6840acaea81deb5c92ccdf2da',1,'PipServices3::Components::Connect::ConnectionParams']]],
   ['getportwithdefault',['GetPortWithDefault',['../class_pip_services3_1_1_components_1_1_connect_1_1_connection_params.html#a466c544ce4f679e9fc8c789819e015c4',1,'PipServices3::Components::Connect::ConnectionParams']]],
   ['getprotocol',['GetProtocol',['../class_pip_services3_1_1_components_1_1_connect_1_1_connection_params.html#af4fd8b940ba7b0a9f8632aedcf86df28',1,'PipServices3::Components::Connect::ConnectionParams']]],
-  ['getprotocolwithdefault',['GetProtocolWithDefault',['../class_pip_services3_1_1_components_1_1_connect_1_1_connection_params.html#a4079f74ef1b87078dce90537b6a663b8',1,'PipServices3::Components::Connect::ConnectionParams']]]
+  ['getprotocolwithdefault',['GetProtocolWithDefault',['../class_pip_services3_1_1_components_1_1_connect_1_1_connection_params.html#a4079f74ef1b87078dce90537b6a663b8',1,'PipServices3::Components::Connect::ConnectionParams']]],
+  ['getvalue',['GetValue',['../class_pip_services3_1_1_components_1_1_state_1_1_state_entry.html#a88c9a3ff369a65ca83d953f6bf1ed9c0',1,'PipServices3::Components::State::StateEntry']]]
 ];