+++ docs/search/classes_2.js
@@ -8,6 +8,8 @@ var searchData=
   ['defaultinfofactory',['DefaultInfoFactory',['../class_pip_services3_1_1_components_1_1_info_1_1_default_info_factory.html',1,'PipServices3::Components::Info']]],
   ['defaultlockfactory',['DefaultLockFactory',['../class_pip_services3_1_1_components_1_1_lock_1_1_default_lock_factory.html',1,'PipServices3::Components::Lock']]],
   ['defaultloggerfactory',['DefaultLoggerFactory',['../class_pip_services3_1_1_components_1_1_log_1_1_default_logger_factory.html',1,'PipServices3::Components::Log']]],
+  ['defaultstatestorefactory',['DefaultStateStoreFactory',['../class_pip_services3_1_1_components_1_1_state_1_1_default_state_store_factory.html',1,'PipServices3::Components::State']]],
+  ['defaulttestfactory',['DefaultTestFactory',['../class_pip_services3_1_1_components_1_1_test_1_1_default_test_factory.html',1,'PipServices3::Components::Test']]],
   ['defaulttracerfactory',['DefaultTracerFactory',['../class_pip_services3_1_1_components_1_1_trace_1_1_default_tracer_factory.html',1,'PipServices3::Components::Trace']]],
   ['diagnosticslogger',['DiagnosticsLogger',['../class_pip_services3_1_1_components_1_1_log_1_1_diagnostics_logger.html',1,'PipServices3::Components::Log']]],
   ['discoveryitem',['DiscoveryItem',['../class_pip_services3_1_1_components_1_1_connect_1_1_memory_discovery_1_1_discovery_item.html',1,'PipServices3::Components::Connect::MemoryDiscovery']]]