+++ docs/search/classes_9.js
@@ -4,5 +4,6 @@ var searchData=
   ['memoryconfigreader',['MemoryConfigReader',['../class_pip_services3_1_1_components_1_1_config_1_1_memory_config_reader.html',1,'PipServices3::Components::Config']]],
   ['memorycredentialstore',['MemoryCredentialStore',['../class_pip_services3_1_1_components_1_1_auth_1_1_memory_credential_store.html',1,'PipServices3::Components::Auth']]],
   ['memorydiscovery',['MemoryDiscovery',['../class_pip_services3_1_1_components_1_1_connect_1_1_memory_discovery.html',1,'PipServices3::Components::Connect']]],
-  ['memorylock',['MemoryLock',['../class_pip_services3_1_1_components_1_1_lock_1_1_memory_lock.html',1,'PipServices3::Components::Lock']]]
+  ['memorylock',['MemoryLock',['../class_pip_services3_1_1_components_1_1_lock_1_1_memory_lock.html',1,'PipServices3::Components::Lock']]],
+  ['memorystatestore',['MemoryStateStore',['../class_pip_services3_1_1_components_1_1_state_1_1_memory_state_store.html',1,'PipServices3::Components::State']]]
 ];