+++ docs/search/classes_d.js
@@ -1,4 +1,6 @@
 var searchData=
 [
-  ['tracetiming',['TraceTiming',['../class_pip_services3_1_1_components_1_1_trace_1_1_trace_timing.html',1,'PipServices3::Components::Trace']]]
+  ['shutdown',['Shutdown',['../class_pip_services3_1_1_components_1_1_test_1_1_shutdown.html',1,'PipServices3::Components::Test']]],
+  ['stateentry',['StateEntry',['../class_pip_services3_1_1_components_1_1_state_1_1_state_entry.html',1,'PipServices3::Components::State']]],
+  ['statevalue',['StateValue',['../class_pip_services3_1_1_components_1_1_state_1_1_state_value.html',1,'PipServices3::Components::State']]]
 ];