+++ docs/search/classes_e.js
@@ -1,4 +1,4 @@
 var searchData=
 [
-  ['yamlconfigreader',['YamlConfigReader',['../class_pip_services3_1_1_components_1_1_config_1_1_yaml_config_reader.html',1,'PipServices3::Components::Config']]]
+  ['tracetiming',['TraceTiming',['../class_pip_services3_1_1_components_1_1_trace_1_1_trace_timing.html',1,'PipServices3::Components::Trace']]]
 ];