+++ docs/search/properties_5.js
@@ -1,4 +1,4 @@
 var searchData=
 [
-  ['key',['Key',['../class_pip_services3_1_1_components_1_1_cache_1_1_cache_entry.html#a57b9bc9aa0215ea5ef21a1bcdcc98d8f',1,'PipServices3::Components::Cache::CacheEntry']]]
+  ['key',['Key',['../class_pip_services3_1_1_components_1_1_cache_1_1_cache_entry.html#a57b9bc9aa0215ea5ef21a1bcdcc98d8f',1,'PipServices3.Components.Cache.CacheEntry.Key()'],['../class_pip_services3_1_1_components_1_1_state_1_1_state_value.html#a83867caf401f5a9552c2aef1b0de3fb4',1,'PipServices3.Components.State.StateValue.Key()']]]
 ];