+++ docs/search/properties_d.js
@@ -1,4 +1,4 @@
 var searchData=
 [
-  ['value',['Value',['../class_pip_services3_1_1_components_1_1_cache_1_1_cache_entry.html#ac5a1e4897fa08665e6d953aa10feae49',1,'PipServices3::Components::Cache::CacheEntry']]]
+  ['value',['Value',['../class_pip_services3_1_1_components_1_1_cache_1_1_cache_entry.html#ac5a1e4897fa08665e6d953aa10feae49',1,'PipServices3.Components.Cache.CacheEntry.Value()'],['../class_pip_services3_1_1_components_1_1_state_1_1_state_value.html#a97d27623761832e9b7a856839f778d4d',1,'PipServices3.Components.State.StateValue.Value()']]]
 ];